World Changes
Added a separate power setting for shooting from the outerworks.  Added
a relay to operate the light (turns on when the launcher is started and
turns off when the StopLauncher Command is called).  Added new auton
modes to test turning ability in auton (turning works when both sides
are set to 100%).  Added a separate "outer intake" talon control (works
like the original intake except with a separate percent voltage: 100%).
Cleaned up the formatting for the subsytems and other class files.